% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_finder.R
\name{job_finder}
\alias{job_finder}
\title{indeed Function}
\usage{
job_finder(query, loc)
}
\arguments{
\item{query:}{The jobtitle or keyword that the user wants to search indeed.com for. The input is not case sensitive
**query is a character string and so it needs to be inputted with quotation marks ""}

\item{loc:}{The location of jobs the user would like to search for. The input is not case sensitive
**loc is a character string as well and so the input needs to be encased around quotation marks. ""}
}
\value{
Table of a dataframe that contains the hiring company, the job title, the description, the location, and the job link 
for the search results of that specific query
}
\description{
The purpose of this function is to take the manual work of searching for jobs out of the hand of the user
The user can enter the type of job they want and the location they would like to work and the function will
pull the most important bits of information about the job and will present them in a compact table for the user
to look at all at once.
This function takes two arguements.
}
\examples{
job_finder("data analyst","Providence")
job_finder("merchandise planner","07002")
job_finder("graphic designer","02912")
}
